+++ TroveTools.NET/Converter/NullValueMultiConverter.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using log4net;
+using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
@@ -13,10 +14,17 @@ namespace TroveTools.NET.Converter
     /// </summary>
     class NullValueMultiConverter : IMultiValueConverter
     {
+        private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+
         public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
         {
-            if (values[0] is string) return string.IsNullOrEmpty(values[0] as string) ? values[1] : values[2];
-            return values[0] == null ? values[1] : values[2];
+            try
+            {
+                if (values[0] is string) return string.IsNullOrEmpty(values[0] as string) ? values[1] : values[2];
+                return values[0] == null ? values[1] : values[2];
+            }
+            catch (Exception ex) { log.Error(string.Format("Error checking for null value (multi converter): [{0}]", values), ex); }
+            return null;
         }
 
         public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)